Giao dien danh sach san pham

+++ QLCuaHangDungCuTheThao/DTO/ChiTietHoaDon.cs
@@ -11,16 +11,16 @@ namespace DTO
         private string maHD;
         private string maSP;
         private int soLuong;
-        private long donGia;
-        private long thanhTien;
+        private int donGia;
+        private int thanhTien;
 
         public string MaHD { get => maHD; set => maHD = value; }
         public string MaSP { get => maSP; set => maSP = value; }
         public int SoLuong { get => soLuong; set { soLuong = value; thanhTien = soLuong * donGia; } }
-        public long DonGia { get => donGia; set => donGia = value; }
-        public long ThanhTien { get => thanhTien;}
+        public int DonGia { get => donGia; set => donGia = value; }
+        public int ThanhTien { get => thanhTien;}
 
-        public ChiTietHoaDon(string maHD, string maSP, int soLuong, long donGia)
+        public ChiTietHoaDon(string maHD, string maSP, int soLuong, int donGia)
         {
             this.MaHD = maHD;
             this.MaSP = maSP;
