+++ QLCuaHangDungCuTheThao/DTO/HoaDon.cs
@@ -3,12 +3,12 @@ public class HoaDon {
     private string maNV;
     private string sdt;
     private string phuongThucThanhToan;
-    private long chietKhau;
+    private int chietKhau;
     private string thoiGianLap;
-    private long tienKhachPhaiTra;
-    private long tienKhachDua;
-    private long tongTienHang;
-    private long traLai;
+    private int tienKhachPhaiTra;
+    private int tienKhachDua;
+    private int tongTienHang;
+    private int traLai;
     private string ghiChu;
     private string trangThai;
 
@@ -16,16 +16,16 @@ public class HoaDon {
     public string MaNV { get => maNV; set => maNV = value; }
     public string SDT { get => sdt; set => sdt = value; }
     public string PhuongThucThanhToan { get => phuongThucThanhToan; set => phuongThucThanhToan = value; }
-    public long ChietKhau { get => chietKhau; set => chietKhau = value; }
+    public int ChietKhau { get => chietKhau; set => chietKhau = value; }
     public string ThoiGianLap { get => thoiGianLap; set => thoiGianLap = value; }
-    public long TienKhachPhaiTra { get => tienKhachPhaiTra; set => tienKhachPhaiTra = value; }
-    public long TienKhachDua { get => tienKhachDua; set => tienKhachDua = value; }
-    public long TongTienHang { get => tongTienHang; set => tongTienHang = value; }
-    public long TraLai { get => traLai; set => traLai = value; }
+    public int TienKhachPhaiTra { get => tienKhachPhaiTra; set => tienKhachPhaiTra = value; }
+    public int TienKhachDua { get => tienKhachDua; set => tienKhachDua = value; }
+    public int TongTienHang { get => tongTienHang; set => tongTienHang = value; }
+    public int TraLai { get => traLai; set => traLai = value; }
     public string GhiChu { get => ghiChu; set => ghiChu = value; }
     public string TrangThai { get => trangThai; set => trangThai = value; }
 
-    public HoaDon(string MaHD, string MaNV, string SDT, string PhuongThucThanhToan, int ChietKhau, string ThoiGianLap, long TienKhachPhaiTra, long TienKhachDua, long TongTienHang, long TraLai, string GhiChu) {
+    public HoaDon(string MaHD, string MaNV, string SDT, string PhuongThucThanhToan, int ChietKhau, string ThoiGianLap, int TienKhachPhaiTra, int TienKhachDua, int TongTienHang, int TraLai, string GhiChu) {
         this.MaHD = MaHD;
         this.MaNV = MaNV;
         this.SDT = SDT;
